requirements.txt
# Django
Django==4.2.7
django-crispy-forms>=2.3
crispy-bootstrap5==2024.10
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Database
psycopg2-binary==2.9.9

# Configuration
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3

# Celery and Task Queue
celery==5.3.4
redis==5.0.1

# Data Acquisition
dataretrieval==1.0.7
requests==2.31.0
pytz==2023.3
python-dateutil==2.8.2
tenacity==8.2.3

# Development/Testing
pytest==7.4.3
pytest-cov==4.1.0
